requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
asgiref==3.11.1
attrs==26.1.0
beautifulsoup4==4.14.3
camelot-py==1.0.9
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
Django==6.0.3
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
fake-useragent==2.2.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
fastar==0.8.0
ghostscript==0.8.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.4.2
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
orjson==3.11.8
outcome==1.3.0.post0
pandas==3.0.0
pdfminer.six==20260107
pillow==12.1.1
playwright==1.58.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
pypdf==5.9.0
pypdfium2==5.4.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
requests==2.33.1
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
selenium==4.43.0
sentry-sdk==2.48.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
sqlparse==0.5.5
starlette==0.50.0
tabulate==0.9.0
trio==0.33.0
trio-websocket==0.12.2
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.12.0
undetected-chromedriver==3.5.5
urllib3==2.6.3
uv==0.9.0
uvicorn==0.40.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.3.2
